+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 1.0.0
+* Initial release.
+* Added CLI tool to generate a static, zero-dependency `licenses.g.dart` file.
+* Supports filtering dependencies by type (`main`, `dev`, `both`).
+* Custom output directory support via `--out`.